Explicit integer

+++ contracts/interfaces/IExchange.sol
@@ -1,8 +1,8 @@
 pragma solidity 0.5.11;
 
 interface IExchange {
-    event Deposited(address payee, uint amount);
-    event Withdrawn(address payee, uint amount);
+    event Deposited(address payee, uint256 amount);
+    event Withdrawn(address payee, uint256 amount);
 
     function deposit(uint256 _amount) external;
     function withdraw(uint256 _amount) external;
